Reland Fixed compiler warning C4267 'conversion from 'size_t' to 'type', possible loss of data' Additional warnings found with more testing and added C4267 warning disable only for angle_libpng
BUG=angleproject:1120

+++ src/libANGLE/Framebuffer.cpp
@@ -274,7 +274,7 @@ bool Framebuffer::hasEnabledColorAttachment() const
 {
     for (size_t colorAttachment = 0; colorAttachment < mData.mColorAttachments.size(); ++colorAttachment)
     {
-        if (isEnabledColorAttachment(colorAttachment))
+        if (isEnabledColorAttachment(static_cast<unsigned int>(colorAttachment)))
         {
             return true;
         }
@@ -297,7 +297,7 @@ bool Framebuffer::usingExtendedDrawBuffers() const
 {
     for (size_t colorAttachment = 1; colorAttachment < mData.mColorAttachments.size(); ++colorAttachment)
     {
-        if (isEnabledColorAttachment(colorAttachment))
+        if (isEnabledColorAttachment(static_cast<unsigned int>(colorAttachment)))
         {
             return true;
         }
